+++ example/main.dart
@@ -3,9 +3,13 @@
 import 'main_caller_logger.dart';
 import 'main_try_handler.dart';
 
-void main() {
+void main() async {
   print(
-      'Run with either `dart example/main_try_catch_handler.dart` or `dart --enable-asserts example/main_try_catch_handler.dart`.');
+      'Run with either `dart example/main.dart` or `dart --enable-asserts example/main.dart`.');
+  print('===== Caller Logger Demo Start =====');
   callerLoggerDemo();
-  tryHandlerDemo();
+  print('===== Caller Logger Demo End =====');
+  print('===== Try Handler Demo Start =====');
+  await tryHandlerDemo();
+  print('===== Try Handler Demo End =====');
 }